+++ lib/constants.dart
@@ -1,0 +1,2 @@
+final int minimumAnagramLength = 1;
+final int maximumAnagramLength = 20;